+++ CHANGES.rst
@@ -4,6 +4,13 @@ Changelog
 Here you can see the full list of changes between each SQLAlchemy-Continuum release.
 
 
+0.3.0 (2013-06-10)
+^^^^^^^^^^^^^^^^^^
+
+- Model changesets
+- Fixed previous and next accessors
+
+
 0.2.1 (2013-06-10)
 ^^^^^^^^^^^^^^^^^^